# backend/src/main/java/com/autospec/controller/ProjectController.java
package com.autospec.controller;

import com.autospec.dto.AgentEventResponse;
import com.autospec.dto.ApproveArtifactResponse;
import com.autospec.dto.ArtifactDiffResponse;
import com.autospec.dto.ArtifactRestoreRequest;
import com.autospec.dto.ArtifactResponse;
import com.autospec.dto.CreateProjectRequest;
import com.autospec.dto.CreateProjectResponse;
import com.autospec.dto.CursorPageResponse;
import com.autospec.dto.CursorPaginationRequest;
import com.autospec.dto.GenerateProjectResponse;
import com.autospec.dto.PaginationRequest;
import com.autospec.dto.ProjectProgressResponse;
import com.autospec.dto.ProjectDashboardItemResponse;
import com.autospec.dto.ProjectResponse;
import com.autospec.dto.RetryTaskResponse;
import com.autospec.dto.ReviewIssueResponse;
import com.autospec.dto.ReviewIssueUpdateRequest;
import com.autospec.dto.ReviewResponse;
import com.autospec.dto.UpdateArtifactRequest;
import com.autospec.entity.AgentEvent;
import com.autospec.entity.AgentTask;
import com.autospec.entity.Artifact;
import com.autospec.entity.Project;
import com.autospec.entity.ReviewIssue;
import com.autospec.service.AgentEventService;
import com.autospec.service.AgentEventStreamService;
import com.autospec.service.AgentOrchestrationService;
import com.autospec.service.ArtifactVersionService;
import com.autospec.service.ArtifactService;
import com.autospec.service.ProjectAccessService;
import com.autospec.service.ProjectDashboardService;
import com.autospec.service.ProjectService;
import com.autospec.service.ReviewIssueService;
import jakarta.validation.Valid;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;

import java.util.List;

@RestController
@RequestMapping("/api/projects")
public class ProjectController {

    private final ProjectService projectService;
    private final AgentOrchestrationService agentOrchestrationService;
    private final ArtifactService artifactService;
    private final ArtifactVersionService artifactVersionService;
    private final ReviewIssueService reviewIssueService;
    private final AgentEventService agentEventService;
    private final AgentEventStreamService agentEventStreamService;
    private final ProjectAccessService projectAccessService;
    private final ProjectDashboardService projectDashboardService;

    public ProjectController(
            ProjectService projectService,
            AgentOrchestrationService agentOrchestrationService,
            ArtifactService artifactService,
            ArtifactVersionService artifactVersionService,
            ReviewIssueService reviewIssueService,
            AgentEventService agentEventService,
            AgentEventStreamService agentEventStreamService,
            ProjectAccessService projectAccessService,
            ProjectDashboardService projectDashboardService
    ) {
        this.projectService = projectService;
        this.agentOrchestrationService = agentOrchestrationService;
        this.artifactService = artifactService;
        this.artifactVersionService = artifactVersionService;
        this.reviewIssueService = reviewIssueService;
        this.agentEventService = agentEventService;
        this.agentEventStreamService = agentEventStreamService;
        this.projectAccessService = projectAccessService;
        this.projectDashboardService = projectDashboardService;
    }

    @GetMapping
    public List<ProjectResponse> projects(
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        Long userId = projectAccessService.resolveUserId(sessionToken);
        return projectAccessService.listVisibleProjects(userId, pagination.limit(), pagination.offset())
                .stream()
                .map(ProjectResponse::from)
                .toList();
    }

    @PostMapping
    public CreateProjectResponse createProject(
            @Valid @RequestBody CreateProjectRequest request,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        Long userId = projectAccessService.resolveUserId(sessionToken);
        Project project = projectService.createProject(request.name(), request.requirement());
        project.setUserId(userId);
        projectService.updateById(project);
        projectAccessService.addOwner(project.getId(), userId);
        return new CreateProjectResponse(project.getId(), project.getStatus());
    }

    @GetMapping("/{projectId}")
    public ProjectResponse project(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        Long userId = projectAccessService.resolveUserId(sessionToken);
        return ProjectResponse.from(projectAccessService.requireVisibleProject(projectId, userId));
    }

    @PostMapping("/{projectId}/generate")
    public GenerateProjectResponse generate(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        Long actorUserId = requireEditor(projectId, sessionToken);
        ProjectProgressResponse progress = agentOrchestrationService.generate(projectId, actorUserId);
        return new GenerateProjectResponse(projectId, progress.status(), progress.percent());
    }

    @PostMapping("/{projectId}/generate-prd")
    public GenerateProjectResponse generatePrd(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        Long actorUserId = requireEditor(projectId, sessionToken);
        ProjectProgressResponse progress = agentOrchestrationService.generatePrd(projectId, actorUserId);
        return new GenerateProjectResponse(projectId, progress.status(), progress.percent());
    }

    @PostMapping("/{projectId}/generate-v4")
    public GenerateProjectResponse generateV4(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestHeader(value = "Idempotency-Key", required = false) String idempotencyKey
    ) {
        Long actorUserId = requireEditor(projectId, sessionToken);
        ProjectProgressResponse progress = agentOrchestrationService.generateV4(projectId, idempotencyKey, actorUserId);
        return new GenerateProjectResponse(projectId, progress.status(), progress.percent());
    }

    @PostMapping("/{projectId}/continue")
    public GenerateProjectResponse continueGeneration(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        Long actorUserId = requireEditor(projectId, sessionToken);
        ProjectProgressResponse progress = agentOrchestrationService.continueAfterApprovedPrd(projectId, actorUserId);
        return new GenerateProjectResponse(projectId, progress.status(), progress.percent());
    }

    @GetMapping("/{projectId}/progress")
    public ProjectProgressResponse progress(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        requireViewer(projectId, sessionToken);
        return agentOrchestrationService.progress(projectId);
    }

    @GetMapping("/{projectId}/artifacts")
    public List<ArtifactResponse> artifacts(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        requireViewer(projectId, sessionToken);
        return artifactService.listByProjectId(projectId, pagination.limit(), pagination.offset())
                .stream()
                .map(ArtifactResponse::from)
                .toList();
    }

    @GetMapping("/dashboard")
    public List<ProjectDashboardItemResponse> dashboard(
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        Long userId = projectAccessService.resolveUserId(sessionToken);
        return projectAccessService.listVisibleProjects(userId, pagination.limit(), pagination.offset())
                .stream()
                .map(projectDashboardService::summarize)
                .toList();
    }

    @GetMapping("/{projectId}/artifacts/page")
    public CursorPageResponse<ArtifactResponse> artifactPage(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(required = false) Integer limit,
            @RequestParam(required = false) String cursor
    ) {
        CursorPaginationRequest pagination = CursorPaginationRequest.of(limit, cursor);
        requireViewer(projectId, sessionToken);
        List<Artifact> fetched = artifactService.listByProjectIdAfterId(
                projectId,
                pagination.afterId(),
                pagination.fetchLimit()
        );
        boolean hasMore = fetched.size() > pagination.limit();
        List<Artifact> page = hasMore
                ? fetched.subList(0, pagination.limit())
                : fetched;
        String nextCursor = hasMore
                ? CursorPaginationRequest.encode(page.get(page.size() - 1).getId())
                : null;
        return new CursorPageResponse<>(
                page.stream().map(ArtifactResponse::from).toList(),
                nextCursor,
                hasMore
        );
    }

    @PutMapping("/{projectId}/artifacts/{artifactId}")
    public ArtifactResponse updateArtifact(
            @PathVariable Long projectId,
            @PathVariable Long artifactId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @Valid @RequestBody UpdateArtifactRequest request
    ) {
        requireEditor(projectId, sessionToken);
        return ArtifactResponse.from(artifactVersionService.updateDraft(
                projectId,
                artifactId,
                request.content(),
                request.expectedLockVersion()
        ));
    }

    @GetMapping("/{projectId}/artifacts/{artifactId}/versions")
    public List<ArtifactResponse> artifactVersions(
            @PathVariable Long projectId,
            @PathVariable Long artifactId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        requireViewer(projectId, sessionToken);
        return artifactVersionService.listVersions(projectId, artifactId, pagination.limit(), pagination.offset())
                .stream()
                .map(ArtifactResponse::from)
                .toList();
    }

    @PostMapping("/{projectId}/artifacts/{artifactId}/approve")
    public ApproveArtifactResponse approveArtifact(
            @PathVariable Long projectId,
            @PathVariable Long artifactId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        requireEditor(projectId, sessionToken);
        Artifact approved = artifactVersionService.approve(projectId, artifactId);
        return new ApproveArtifactResponse(approved.getId(), approved.getStatus(), approved.getVersion());
    }

    @GetMapping("/{projectId}/artifacts/{artifactId}/diff")
    public ArtifactDiffResponse artifactDiff(
            @PathVariable Long projectId,
            @PathVariable Long artifactId,
            @RequestParam Long againstArtifactId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        requireViewer(projectId, sessionToken);
        return artifactVersionService.diff(projectId, againstArtifactId, artifactId);
    }

    @PostMapping("/{projectId}/artifacts/{artifactId}/restore")
    public ArtifactResponse restoreArtifact(
            @PathVariable Long projectId,
            @PathVariable Long artifactId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @Valid @RequestBody ArtifactRestoreRequest request
    ) {
        requireEditor(projectId, sessionToken);
        return ArtifactResponse.from(artifactVersionService.restore(
                projectId,
                artifactId,
                request.expectedLatestLockVersion()
        ));
    }

    @GetMapping("/{projectId}/review")
    public ReviewResponse review(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        requireViewer(projectId, sessionToken);
        List<ReviewIssueResponse> issues = reviewIssueService.listByProjectId(projectId, pagination.limit(), pagination.offset())
                .stream()
                .map(this::reviewIssueResponse)
                .toList();
        return new ReviewResponse(agentOrchestrationService.reviewScore(projectId), issues);
    }

    @PutMapping("/{projectId}/review/issues/{issueId}")
    public ReviewIssueResponse updateReviewIssue(
            @PathVariable Long projectId,
            @PathVariable Long issueId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @Valid @RequestBody ReviewIssueUpdateRequest request
    ) {
        Long actorUserId = requireEditor(projectId, sessionToken);
        return reviewIssueResponse(reviewIssueService.updateDisposition(
                projectId,
                issueId,
                request.status(),
                request.resolution(),
                request.resolvedInArtifactId(),
                actorUserId
        ));
    }

    @GetMapping("/{projectId}/events/history")
    public List<AgentEventResponse> eventHistory(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(defaultValue = "50") Integer limit,
            @RequestParam(defaultValue = "0") Integer offset
    ) {
        PaginationRequest pagination = PaginationRequest.of(limit, offset);
        requireViewer(projectId, sessionToken);
        return agentEventService.listByProjectId(projectId, pagination.limit(), pagination.offset())
                .stream()
                .map(AgentEventResponse::from)
                .toList();
    }

    @GetMapping("/{projectId}/events/history/page")
    public CursorPageResponse<AgentEventResponse> eventHistoryPage(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken,
            @RequestParam(required = false) Integer limit,
            @RequestParam(required = false) String cursor
    ) {
        CursorPaginationRequest pagination = CursorPaginationRequest.of(limit, cursor);
        requireViewer(projectId, sessionToken);
        List<AgentEvent> fetched = agentEventService.listByProjectIdAfterId(
                projectId,
                pagination.afterId(),
                pagination.fetchLimit()
        );
        boolean hasMore = fetched.size() > pagination.limit();
        List<AgentEvent> page = hasMore
                ? fetched.subList(0, pagination.limit())
                : fetched;
        String nextCursor = hasMore
                ? CursorPaginationRequest.encode(page.get(page.size() - 1).getId())
                : null;
        return new CursorPageResponse<>(
                page.stream().map(AgentEventResponse::from).toList(),
                nextCursor,
                hasMore
        );
    }

    @GetMapping(value = "/{projectId}/events", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
    public SseEmitter streamEvents(
            @PathVariable Long projectId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        requireViewer(projectId, sessionToken);
        return agentEventStreamService.subscribe(projectId);
    }

    @PostMapping("/{projectId}/tasks/{taskId}/retry")
    public RetryTaskResponse retryTask(
            @PathVariable Long projectId,
            @PathVariable Long taskId,
            @RequestHeader(value = "X-AutoSpec-Session-Token", required = false) String sessionToken
    ) {
        requireEditor(projectId, sessionToken);
        AgentTask retry = agentOrchestrationService.retryTask(projectId, taskId);
        return RetryTaskResponse.from(retry);
    }

    private void requireViewer(Long projectId, String sessionToken) {
        projectAccessService.requireProjectRole(
                projectId,
                projectAccessService.resolveUserId(sessionToken),
                "OWNER",
                "EDITOR",
                "VIEWER"
        );
    }

    private Long requireEditor(Long projectId, String sessionToken) {
        Long userId = projectAccessService.resolveUserId(sessionToken);
        projectAccessService.requireProjectRole(
                projectId,
                userId,
                "OWNER",
                "EDITOR"
        );
        return userId;
    }

    private ReviewIssueResponse reviewIssueResponse(ReviewIssue issue) {
        return new ReviewIssueResponse(
                issue.getId(),
                issue.getIssueKey(),
                issue.getSeverity(),
                issue.getIssueType(),
                issue.getArtifactType(),
                issue.getArtifactPath(),
                issue.getRequirementId(),
                issue.getDescription(),
                issue.getSuggestion(),
                issue.getEvidence(),
                issue.getStatus(),
                issue.getOwnerUserId(),
                issue.getResolution(),
                issue.getResolvedInArtifactId()
        );
    }
}