# backend/src/main/resources/contracts/autospec-backend-v1.openapi.yaml
openapi: 3.0.3
info:
  title: AutoSpec Backend API
  version: v1
  description: Spring Boot API contract for AutoSpec project orchestration, artifacts, exports, and observability.
servers:
  - url: /
security:
  - SessionToken: []
paths:
...
  /api/projects:
    get:
      summary: List projects visible to the session user.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Project history visible to the session user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a project from a natural-language requirement.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Created project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/projects/dashboard:
    get:
      summary: List project dashboard summaries visible to the session user.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Project dashboard summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDashboardItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/projects/{projectId}:
    get:
      summary: Fetch a project visible to the session user.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Project detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{projectId}/generate:
    post:
      summary: Run the baseline multi-agent project generation workflow.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Generation progress snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/generate-prd:
    post:
      summary: Generate the PRD stage for a project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: PRD generation progress snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/generate-v4:
    post:
      summary: Run the V4 multi-agent workflow with idempotency.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Generation progress snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/projects/{projectId}/continue:
    post:
      summary: Continue generation after PRD approval.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Continued generation progress snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/projects/{projectId}/progress:
    get:
      summary: Fetch current project generation progress.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Current project generation progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectProgressResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/projects/{projectId}/events:
    get:
      summary: Stream live Agent events for a project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Server-sent Agent event stream.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/events/history:
    get:
      summary: List persisted Agent node events.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Agent event history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/events/history/page:
    get:
      summary: List persisted Agent events by stable ascending ID cursor.
      description: The cursor resumes after the previous page's last event ID and is unaffected by later inserts.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationCursor'
      responses:
        '200':
          description: Cursor page of Agent events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentEventCursorPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/projects/{projectId}/review:
    get:
      summary: Fetch the review score and paged review issues.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Review score and issues.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/tasks/{taskId}/retry:
    post:
      summary: Retry a failed Agent task.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Retried Agent task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetryTaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/projects/{projectId}/artifacts:
    get:
      summary: List project artifacts.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Artifact list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtifactResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/review/issues/{issueId}:
    put:
      summary: Start, resolve, reopen, or explicitly ignore a first-class review issue.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIssueUpdateRequest'
      responses:
        '200':
          description: Updated review issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewIssueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{projectId}/artifacts/page:
    get:
      summary: List project artifacts by stable ascending ID cursor.
      description: New rows are returned only after the previous page's last ID, so inserts do not shift existing pages.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationCursor'
      responses:
        '200':
          description: Cursor page of project artifacts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactCursorPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/artifacts/{artifactId}:
    put:
      summary: Update an artifact by creating a new draft version.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtifactRequest'
      responses:
        '200':
          description: Updated artifact draft version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
  /api/projects/{projectId}/artifacts/{artifactId}/versions:
    get:
      summary: List artifact version history.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ArtifactId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Artifact version history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtifactResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    SessionToken:
      type: apiKey
      in: cookie
      name: AUTOSPEC_SESSION
  parameters:
    SessionToken:
      name: AUTOSPEC_SESSION
      in: cookie
      required: true
      schema:
        type: string
...
    PaginationLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    PaginationOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name, requirement]
      properties:
        name:
          type: string
          minLength: 1
        requirement:
          type: string
          minLength: 1
    CreateProjectResponse:
      type: object
      required: [projectId, status]
      properties:
        projectId:
          type: integer
          format: int64
        status:
          type: string
    ProjectResponse:
      type: object
      required: [projectId, name, originalRequirement, status, createdAt, updatedAt]
      properties:
        projectId:
          type: integer
          format: int64
        name:
          type: string
        originalRequirement:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...